Replace qt4-automoc with qt4_wrap_cpp
