+++ libavogadro/src/atom.cpp
@@ -293,5 +293,3 @@ using Eigen::Vector3d;
    }
 
  } // End namespace Avogadro
-
-#include "atom.moc"
