+++ libavogadro/src/cube.cpp
@@ -322,5 +322,3 @@ namespace Avogadro {
   }
 
 } // End namespace Avogadro
-
-#include "cube.moc"
