+++ libavogadro/src/engine.cpp
@@ -377,5 +377,3 @@ namespace Avogadro {
       return m_molecule->bonds();
   }
 }
-
-#include "engine.moc"
