+++ libavogadro/src/engines/bsdyengine.cpp
@@ -442,7 +442,5 @@ namespace Avogadro
   }
 
 }
-
-#include "bsdyengine.moc"
 // This is a static engine...
 // Q_EXPORT_PLUGIN2( bsdyengine, Avogadro::BSDYEngineFactory )
