+++ libavogadro/src/extension.cpp
@@ -80,5 +80,3 @@ namespace Avogadro {
   }
 
 }
-
-#include "extension.moc"
