+++ libavogadro/src/pluginmanager.cpp
@@ -784,5 +784,3 @@ namespace Avogadro {
   }
 
 }
-
-#include "pluginmanager.moc"
