+++ libavogadro/src/primitive.cpp
@@ -77,5 +77,3 @@ namespace Avogadro {
   }
 
 }
-
-#include "primitive.moc"
