+++ libavogadro/src/pythonerror.cpp
@@ -112,4 +112,3 @@ namespace Avogadro
 
 } // namespace
 
-#include "pythonerror.moc"
