+++ libavogadro/src/pythonextension_p.cpp
@@ -340,5 +340,3 @@ namespace Avogadro
   }
 
 }
-
-#include "pythonextension_p.moc"
