+++ libavogadro/src/toolgroup.cpp
@@ -206,5 +206,3 @@ namespace Avogadro {
   }
 
 } // end namespace Avogadro
-
-#include "toolgroup.moc"
